Suppress vendor prompt banner entirely; fix leaked render hooks on both panels
