+++ .cleanup-tasks/2026_08_26_desktop_notifications_feature.md
@@ -28,7 +28,8 @@ deploy cycle.
 
 ## Additional Cleanup
 
-- The following file branches on `DesktopNotificationsFeature::active()` — keep the active branch and delete the inactive branch in each:
+- The following file branches on `DesktopNotificationsFeature::active()` — keep the active branch and delete the inactive branch:
     - `app/Filament/Pages/ManageBrowserNotifications.php` (`canAccess()`)
-    - `app/Filament/Plugins/GatedBrowserNotificationsPlugin.php` (`register()` prompt banner hook) — once inactive branch removed, delete this class entirely and register `Emuniq\FilamentBrowserNotifications\BrowserNotificationsPlugin::make()` directly in `AdminPanelProvider` instead.
 - Remove the activation of `DesktopNotificationsFeature` from `database/migrations/2026_08_25_130000_create_push_subscriptions_table.php` when deleting the flag class.
+
+`app/Filament/Plugins/HeadOnlyBrowserNotificationsPlugin.php` is unrelated to this flag — it permanently suppresses the vendor's floating prompt banner in favor of the Notifications page card, and stays regardless of this cleanup.
